ml-service/requirements.txt
# ML Service Dependencies
# Python 3.11+ required

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Machine Learning
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
scipy==1.12.0

# MySQL Connector
mysql-connector-python==8.3.0
pymysql==1.1.0

# Caching
redis==5.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Logging
loguru==0.7.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
